+++ tests/distributed/test_comm_ops.py
@@ -10,7 +10,7 @@ from vllm.model_executor.parallel_utils.communication_op import (
     tensor_model_parallel_all_reduce,
     tensor_model_parallel_all_gather,
 )
-from tests.distributed.comm_utils import init_test_distributed_environment, multi_process_tensor_parallel
+from vllm.test_utils import init_test_distributed_environment, multi_process_tensor_parallel
 
 
 @ray.remote(num_gpus=1, max_calls=1)
